refactor: Simplify routes and handlers
